inserção de reserva de horarios

+++ src/server/database/bancoDeDados/providers/barbeiro/deleteBarbeiro.js
@@ -1,7 +1,13 @@
 const{ StatusCodes } = require('http-status-codes');
 const knex = require('../../database');
-const deleteBarbeiro = async (id)=>{
+const deleteBarbeiro = async (id,role)=>{
     try{
+        if(role !== 'barbeiro'){
+            return {
+                status:StatusCodes.FORBIDDEN,
+                message:'você não tem permisão'
+            }
+        }
         const resultado = await knex('barbeiros').where('id',id).delete()
         if(resultado){
             return {message:'barbeiro deletado',status:StatusCodes.OK }
